move rename command to archiver.rename
